Creando archivo de parámetros para despiegue y cambiando tipo de dato para comparar con cantidad de transacciones permitidas por la cuenta

+++ src/main/java/com/bootcamp/personal/passiveaccounts/entity/AccountType.java
@@ -21,7 +21,7 @@ public class AccountType {
     private String abbreviation;
     private String isoCurrencyCode;
     private BigDecimal interesRate;
-    private Integer transactionsNumber;
+    private Long transactionsNumber;
     private Double mainteanceCost;
     private short registrationStatus;
     private Date insertionDate;
